[Metaschedule, Refactor] Move MultiLevelTilingNode decl to a header
